refactor: :art: change folder structure and authentication service
