+++ utils/jwt_auth_users.py
@@ -1,0 +1,60 @@
+from fastapi import Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordBearer
+from jose import jwt, JWTError
+from passlib.context import CryptContext
+from database.db import users_db
+from models.user import User, UserDB
+import os
+
+from dotenv import load_dotenv
+
+load_dotenv()
+
+
+ALGORITHM = os.getenv("ALGORITHM")
+ACCESS_TOKEN_DURATION = int(os.getenv("ACCESS_TOKEN_DURATION"))
+SECRET = os.getenv("SECRET")
+
+
+crypt = CryptContext(schemes=["bcrypt"])
+
+oauth2 = OAuth2PasswordBearer(tokenUrl="login")
+
+exception = HTTPException(
+    status_code=status.HTTP_400_BAD_REQUEST,
+    detail="Credenciales de Autenticación Invalidas",
+    headers={"WWW.Authenticate": "Bearer"},
+)
+
+
+def search_user_db(username: str):
+    if username in users_db:
+        return UserDB(**users_db[username])
+
+
+def search_user(username: str):
+    if username in users_db:
+        return User(**users_db[username])  # retornar users del arreglo
+
+
+# Autenticacion de usuario encriptado
+async def auth_user(token: str = Depends(oauth2)):
+    try:
+        username = jwt.decode(token, SECRET, algorithms=[ALGORITHM]).get("sub")
+        if username is None:
+            raise exception
+
+    except JWTError:
+        raise exception
+
+    return search_user(username)
+
+
+# si el user esta inactivo= disabled mandar mensaje
+async def current_user(user: User = Depends(auth_user)):
+    if user.disabled:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST, detail="Usuario inactivo"
+        )
+
+    return user
